+++ R/evaluate_measurement_model.R
@@ -21,7 +21,9 @@ validity <- function(seminr_model) {
     htmt            = HTMT(seminr_model),
     cross_loadings  = cross_loadings(seminr_model),
     item_vifs       = item_vifs(seminr_model),
-    antecedent_vifs = antecedent_vifs(seminr_model)
+    # TODO: consider if antecedent vifs should be part of structural results
+    antecedent_vifs = antecedent_vifs(
+      seminr_model$smMatrix, stats::cor(seminr_model$construct_scores))
   )
 }